Step 3:  Alter all source to be TypeScript
This changes all files in the `src` directory to be TypeScript files, and alters the launching JS file, `bin/run.js` to execute node with a custom loader from `ts-node` that will transpile the TypeScript ESM Modules into Node ESM Modules at the start.  This also adds the `ts-node` package as a dependency.