Step 4:  Add runtime type checking for Person and Address instances.
This introduces two "special" types of functions in TypeScript:  Type Guards and Type Assertions.  This allows us to inspect the data at runtime to "narrow" the type of a given reference to a known, described type.  This allows the compiler to continue doing static analysis _past_ the point of type guarding / assertion while being confident the referenced variable meets thet shape we have declared.  If it does not, the runtime can throw a clear AssertionError (or any other Error) to describe the situation.

Type Guards & Assertions are incredibly powerful for dealing with unstructured data, or data whose structure may change over time.